% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\docType{data}
\name{fail_fiber}
\alias{fail_fiber}
\title{The failure stresses (in GPA) of single carbon fibers}
\format{
A dataframe with 65 observations.
}
\usage{
data(fail_fiber, package = "WL")
}
\description{
The data represents the strength measured in GPA for single carbon fibers and impregnated
1000-carbon fiber tows.
Single fibers were tested under tension at gauge lengths of 50 mm,
with sample sizes n = 65 (Bader et al., 1982).
}
\references{
Bader, M. G., Priest, A. M. (1982). Statistical aspects of fibre and bundle strength in
hybrid composites. In: Hayashi, T., Kawata, K., Umekawa, S., eds. Progress in Science and
Engineering Composites: Proceedings of the Fourth International Conference on Composite Materials,
ICCM-IV. Tokyo: Japan Society for Composite Materials.

Hyoung-Moon Kim. and Yu-Hyeong Jang. (2020). New Closed-Form Estimators for Weighted Lindley
Distribution. \emph{ }, submitted.
}
\keyword{datasets}
