% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\docType{data}
\name{fail_time}
\alias{fail_time}
\title{The lifetime failure data of an electronic device}
\format{
A dataframe with 18 observations.
}
\usage{
data(fail_time, package = "WL")
}
\description{
It represents the lifetime failure data of an electronic device (Wang, 2000).
}
\references{
Wang, F. (2000). A new model with bathtub-shaped failure rate using an additive Burr XII
distribution. \emph{Reliability Engineering and System Safety}, 70:305-312.

Hyoung-Moon Kim. and Yu-Hyeong Jang. (2020). New Closed-Form Estimators for Weighted Lindley
Distribution. \emph{ }, submitted.
}
\keyword{datasets}
