% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\docType{data}
\name{lifetime_alum}
\alias{lifetime_alum}
\title{The lifetimes of aluminum specimens}
\format{
A dataframe with 102 observations.
}
\usage{
data(lifetime_alum, package = "WL")
}
\description{
This data set represents the lifetimes of 102 aluminum specimens exposed to
a maximum stress/cycle of 26000 PSI (PSI is a unit of pressure expressed in
pounds of force per square inch of area; it stands for Pounds per Square Inch.
1 PSI = 6894 Pascals = 0.070 atmospheres = 51.715 torr) (Owen et al., 2000).
}
\references{
Owen W.J., Padget W.J. (2000). A Birnbaum-Saunders accelerated life model.
\emph{IEEE Trans Reliability}, 49(2):224-229.

Hyoung-Moon Kim. and Yu-Hyeong Jang. (2020). New Closed-Form Estimators for Weighted Lindley
Distribution. \emph{ }, submitted.
}
\keyword{datasets}
