% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\name{summary.WL}
\alias{summary.WL}
\alias{print.summary.WL}
\title{Summarizing WL function}
\usage{
\method{summary}{WL}(object, ...)

\method{print}{summary.WL}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class "WL" made by the function \code{WL}.}

\item{...}{not used, but exists because of the compatibility.}

\item{x}{an object of class "summary.WL".}

\item{digits}{a numeric number of significant digits.}
}
\description{
\code{summary} method for a class "WL".
}
\examples{
example <- fail_fiber
result <- WL(example, est_method="MLEc")
summary(result)

}
